Source_submit/libs/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
unidecode>=1.3.0
tqdm>=4.62.0
accelerate>=0.20.0
datasets>=2.12.0
timm==1.0.14
tensorflow==2.18.0
